renamed files and created QuestionEditor
